Rename 'product.dart' to 'receipt_product.dart'
