+++ bin/receipt_product.dart
